requirements.txt
faiss-cpu
pandas
numpy
Flask
sentence-transformers
langchain
